+++ src/enum/function.enum.ts
@@ -1,0 +1,3 @@
+export enum Function {
+  TIME_SERIES_INTRADAY = 'TIME_SERIES_INTRADAY',
+}