+++ src/enum/interval.enum.ts
@@ -1,0 +1,7 @@
+export enum Interval {
+  ONE_MIN = '1min',
+  FIVE_MIN = '5min',
+  FIFTEEN_MIN = '15min',
+  THIRTY_MIN = '30min',
+  SIXTY_MIN = '60min',
+}